Refactor some properties shared by all server-side call contexts into a common `ServerCallContext` protocol, while renaming the existing `ServerCallContext` class to `ServerCallContextBase`.
Motivation: As opposed to all three other call types, unary calls are not passed an instance of the previous `ServerCallContext` class, but rather of the `StatusOnlyCallContext` protocol. If one wanted to create a method that uses only a call's event loop and request headers but wants to process all *four* types of calls, that was not possible before. With this change, such a method could simply accept an instance of `ServerCallContext`, which the call arguments provided to all four methods now conform to.