Disable iOS 12 cellular connectivity monitor by default
We've started seeing some crashes when using the `CTServiceRadioAccessTechnologyDidChange` notification within `ClientNetworkMonitor ` in production on iOS 12. This is an issue that's been reported in this radar: https://openradar.appspot.com/46873673

To provide a workaround for this problem, I'm adding the option to opt out of the new iOS 12 API and to instead use the stable pre-iOS 12 API.

I'm planning to release a new version of SwiftGRPC with this fix after it's merged.