Rename GRPCNIO to PlatformSupport
Motivation:

We were never really satisfied with the name "GRPCNIO" for the
namespace with functions to provide event loop groups
and bootstraps for NIO and NIOTS in a platform-agnostic way.

Modifications:

Rename GRPCNIO to PlatformSupport, add an availability attribute
to GRPCNIO to make it easier for users to update their code after
the renaming.

Result:

Slightly better naming.