Bump required version of NIOSSL (#559)
Motivation:

Prior to NIOSSL 2.4 it was possible to silently drop some certificates
from a PEM bundle when loading from a file or buffer. These APIs were
deprecated in 2.4 and our usage of them was causing compiler warnings.

Modifications:

Set the minimum version of NIOSSL to 2.4; update usage to the new APIs.

Result:

Fewer compiler warnings.