Replace Commander usage with our own CLI parsing
Motivation:

We're relying on third-party code for parsing CLI arguments, however,
this is only being used for interop tests and the echo example and isn't
required to use the GRPC module. However when another package depends on
GRPC, SwiftPM will also bring in the third-party code (and all of its
transitive dependencies) even if no targets used by the other package
rely on that code.

Modifications:

Remove usages of "Commander" and replace with some basic CLI parsing.

Result:

Our CLIs are less colorful, but we drop a dependency.