Add HelloWorld example
Motivation:

To help users get started quickly with gRPC Swift it would be useful to
have a "hello world" example. grpc.io has such examples for
a number of languages: https://grpc.io/docs/quickstart/ but not for Swift.

Another motivating factor for this change was feedback from the gRPC on
the SSWG draft proposal suggesting that it would be useful to frame the
proposal from a canonical example, either "HelloWorld" or "RouteGuide".

Modifications:

- Add a "HelloWorld" example
- Create a "quick-start" guide modelled on those on grpc.io

Result:

- No functional changes
- Another example and tutorial