Remove unexported executables from products (#592)
Motivation:

From https://developer.apple.com/documentation/swift_packages/product

> Executable
> Use an executable product to vend an executable target.
> Use this only if you want to make the executable available to clients.

Several executable products are not meant to be made available to
clients, but rather exist for internal integration or performance tests.
Further, most of these products are not namespaced and have generic names.
The code-generator product, `protoc-gen-grpc-swift`, appears to be the
only executable intended for export and it is already appropriately
namespaced.

Modifications:

Unexport the executable products that are not intended for use by clients.
Change the `Makefile` to reference the executable targets rather than their
derived products.

Result:

Fewer potential namespace conflicts. This fixed #591.