Use "swift run" for interop tests in CI
Motivation:

60205adb (#592) cleared up the product namespace by removing executables
which were not meant to be exposed to clients. However the post-merge CI
failed (i.e. interop tests) because the executable targets no longer
exist. This was missed because the CI for the PR was still able to build
the targets and because the products had different names to the targets.

Modifications:

Use 'swift run' in the CI script used after merging PRs.

Result:

Interop tests should be runnable when a PR is merged.