Fix parsing of content-type on the server.
Motivation:

The server determines content-type by getting the first header with the
content-type key, and then attempts to match it to a ContentType enum
case. If the content-type header was missing the value was assumed to be
"application/grpc", however if the value was present but did not match
any of the ContentType enum cases then the content type would not be
set.

Modifications:

- Fallback to using "application/grpc" if the content-type is present
  but does not match one of the cases (e.g. "application/grpc+proto",
  which is valid).

While this modification isn't strictly correct, (see
https://github.com/grpc/grpc-swift/issues/581) it matches the documented
behaviour.

Result:

The server will assume that content-type is "application/grpc" if the
content-type is not correct or missing.