Update the client to use the new state machine backed channel handler
Motivation:

We recently rewrote the client channel handling pipeline to be backed by
a single state machine which also does not require the conversion from
HTTP2 to HTTP1. This PR update the client to make use of this new
pipeline.

Modifications:

- Update the client request and response channel handlers as they now
  use different types and no longer require their own state tracking
  since it is provided by the GRPCClientChannelHandler
- Update the different call types and base client call to make use of
  these new types.
- Delete unused channel handlers.
- Updated tests to make use of new types.
- Fixed a bug in one of the interop test cases.

Result:

The client now uses our new channel handler and state machine and no
longer requires the conversion from HTTP2 to HTTP1.