Allow client error delegate to have a logger injected
Motivation:

When errors are logged it's useful to have some contextual information.
We can enable this by injecting a `logger` into the error delegate.

Modifications:

Change the API for the `ClientErrorDelegate` to also pass in a logger.
Remove additional logging from next to the call to the delegate.

Result:

Users can log errors with more contextual information in the delegate.