Avoid protocol check after TLS handshake (#685)
Motivation:

When tls handshake is over there is no need to raise a failure if no
protocol has been negotiated.

Modifications:

The ```TLSVerificationHandler.userInboundEventTriggered``` method has
been updated so that it always succeed. However it will notice when a
protocol could not be negotiated

Result:

This change allows a successful connection even when a protocol is not
negotiated. By doing so we are more consistent with other grpc
implementations.