Use metadata in log messages
Motivation:

I was analyzing some logs produced by grpc-swift but some of
the messages lack structure, making it more difficult than it
should be to run queries on the logs.

SwiftLog supports structured logging and its API allow us to
log a message and some metadata associated to it.
Keeping the message string constant and logging the variable
pieces of information as metadata, the logs produced by grpc-swift
are going to be more easily analysed and processed.

Modifications:

This commit changes the logging calls where we had a variable
message to have a constant message. The variable pieces of information
are stored as metadata.

Result:

grpc-swift will produce log messages that can be analyzed and processed
more easily.