Rename generated clients from FooServiceClient to FooClient
Motivation:

Issue #735 highlighted that the naming of generated clients is clunky.
Let's improve it before the API is locked down.

Note: it's possible to create a shim to create fix-its for this by creating a
typealias from the old name to the new name, and then deprecating the
typealias:

```swift
@available(*, deprecated, renamed: "Echo_EchoClient")
typealias Echo_EchoServiceClient = Echo_EchoClient
```

Modifications:

- Update codegen
- Regenerate clients and update usages

Result:

Generate clients have better names