Allow users to choose a policy for starting RPCs
Motivation:

In most cases RPCs should be started once a connection has become
'ready', this allows RPCs to wait for transient network failures to be
recovered from instead of just failing. However in some contexts this is
undesirable, and failing RPCs quickly when it does not have a connection
is the right thing to do.

Modifications:

- Add a 'CallStartPolicy' which allows users RPCs to configure their
  connection such that RPCs will wait for the connection to be ready
  before starting the RPC or fail quickly if the most recent connection
  attempt failed.

Result:

Users can fail RPCs quickly if they do not wish to wait for a connection
to be eventually established.