Provide a 'FakeChannel'
Motivation:

When consuming gRPC it is often helpful to be able to write tests that
ensure the client is integrated correctly. At the moment this is only
possible by running a local gRPC server with a custom service handler to
return the responses you would like to test.

Modifications:

This is a continuation of the work started in #855.

- This change addes a 'FakeChannel' this is the glue that binds the Call
  objects with the 'fake responses' added in the aforementioned pull
  request.
- Also adds a 'write capturing' handler which forwards request parts to
  a handler provided on the fake response.
- Appropriate internal initializers on each of the call types.

Result:

- Users can manually create 'test' RPCs.