Alter the documented requirements of `MessageSerializer`
Motivation:

requirements that implementations must prefix their serialized messages
with a compression byte and `UInt32` length. The motivation behind this
was to avoid a reallocation later on when the message was being framed.
The result was that in some cases the framing would be done by the
serialier and in other cases by the framer (or
`LengthPrefixedMessageWriter`).

Modifications:

- Remove the requirement for the serialier to frame the message
- Optimize the `LengthPrefixedMessageWriter` so that it checks whether
  there is enough space _before_ the serialized bytes in which the
  compression byte and length could be written.
- gRPC-Web: Replace the single response buffer with a circular buffer of
  buffers; this avoids writing into a buffer and simplifies the
  `LengthPrefixedMessageWriter`.

Result:

- Better separation of concerns.