Don't specify a custom module name when generating Podspecs
Motivation:

The "gRPC-Core" cocoapod specifies its module name as "grpc"; this is
unfortunate because we also specify our module name as "GRPC". This is
problematic of both are used in the same dependency graph. In CocoaPods
`module_name` is "name to use for the framework / clang module which
will be generated". We therefore end up with two "GRPC" frameworks.

If the module name is not specified it will default to the spec name
instead.

Modifications:

- Don't generate a custom module name when generating Podspecs
- Delete trailing whitespace

Result:

- Resolves #905
- Generated frameworks match Podspec names (i.e. "gRPC-Swift" gets
  converted to "gRPC_Swift")