Switch a `numericCast` to the required type
Motivation:

The server codec uses `numericCast` to turn an `Int` into `_Capacity` in
order to call `ByteBuffer.clear(minimumCapacity:)` which accidentally
requires an underscore-internal type for capacity. The as yet unreleased
fix in NIO deprecates that function and adds an identical signature
using `Int` instead of `_Capacity`.

Using `numericCast` in this instance will result in a compiler error
since the type is ambiguous.

Modifications:

- Replace the `numericCast` with `UInt32`, the alias of `_Capacity`.

Result:

- When the NIO fix is released, gRPC will still be able to compile.